changes to benchmark: - created the competitive AIAO Interrupted benchmark
Problem: The stxxl_pq throws assertion errors when validation is done from the benchmark method itself (using the same instance of the stxxl pq). The validation procedure works perfectly fine (as in the mlrpq only benchmarks), but trying to validate the results using the stxxl pq after the stxxl pq was benchmarked does not work. Therefore the validation procedure is commented out for now.